Create unit-tests for Player class; refactor Deck.ts
